fix(entities): resolve hibernate schema validation errors - adjust column types and mappings for multiple entities to match DB schema - add missing @Entity annotations and fix composite keys - examples: Actor.actorId (SMALLINT UNSIGNED), Film.releaseYear (Year), Staff.picture (BLOB)

+++ src/main/java/org/example/safargulov/projecthibernate2/dto/FilmDto.java
@@ -16,7 +16,7 @@ public class FilmDto {
     private String title;
     private String description;
     private LocalDateTime releaseYear;
-    private Integer length;
+    private Short length;
     private Set<SpecialFeature> specialFeatures;
     private LanguageDto language;
     private LanguageDto originalLanguage;
