+++ src/main/java/org/example/safargulov/projecthibernate2/entity/FilmText.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class FilmText {
     @Id
     @Column(name = "film_id", nullable = false)
-    private Integer filmId;
+    private Short filmId;
 
     @Column(name = "title", nullable = false)
     private String title;
